requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
PyJWT
Werkzeug
python-dotenv
flasgger